Add confirm delete alert dialog

+++ lib/widgets/tasks_list.dart
@@ -18,7 +18,37 @@ class TasksList extends StatelessWidget {
                 taskData.updateTask(task);
               },
               longPressCallback: () {
-                taskData.deleteTask(task);
+                return showDialog<void>(
+                  context: context,
+                  barrierDismissible: false, // user must tap button!
+                  builder: (BuildContext context) {
+                    return AlertDialog(
+                      title: Text('Confirm Delete?'),
+                      content: SingleChildScrollView(
+                        child: ListBody(
+                          children: <Widget>[
+                            Text('Would you like to delete this task?'),
+                          ],
+                        ),
+                      ),
+                      actions: <Widget>[
+                        TextButton(
+                          child: Text('Cancel'),
+                          onPressed: () {
+                            Navigator.of(context).pop();
+                          },
+                        ),
+                        TextButton(
+                          child: Text('Delete'),
+                          onPressed: () {
+                            taskData.deleteTask(task);
+                            Navigator.of(context).pop();
+                          },
+                        ),
+                      ],
+                    );
+                  },
+                );
               },
             );
           },
